[`pylint`] Demote `PLW1510` fix to display-only
Summary
--

Closes #17091. `PLW1510` checks for `subprocess.run` calls without a `check`
keyword argument and previously had a safe fix to add `check=False`. That's the
default value, so technically it preserved the code's behavior, but as discussed
in #17091 and #17087, Ruff can't actually know what the author intended.

I don't think it hurts to keep this as a display-only fix instead of removing it
entirely, but it definitely shouldn't be safe at the very least.

Test Plan
--

Existing tests